back/requirements.txt
fastapi
uvicorn
requests
pydantic
pytest
neo4j
sse-starlette
python-dotenv
